Embed refcodes: persist the url with refcodes to the block
