+++ actblue-contributions/blocks/custom/actblue-embed/save.js
@@ -7,9 +7,10 @@ import classnames from "classnames/dedupe";
  * WordPress dependencies
  */
 import { RichText } from "@wordpress/block-editor";
+import urlWithQueryConfiguration from "./urlWithQueryConfiguration";
 
 const EmbedSave = ({ attributes }) => {
-	const { url, caption, type } = attributes;
+	const { url, caption, type, refcode } = attributes;
 
 	if (!url) {
 		return null;
@@ -22,7 +23,12 @@ const EmbedSave = ({ attributes }) => {
 	return (
 		<figure className={embedClassName}>
 			<div className="wp-block-embed__wrapper">
-				{`\n${url}\n` /* URL needs to be on its own line. */}
+				{
+					`\n${urlWithQueryConfiguration({
+						url,
+						refcode,
+					})}\n` /* URL needs to be on its own line. */
+				}
 			</div>
 			{!RichText.isEmpty(caption) && (
 				<RichText.Content tagName="figcaption" value={caption} />
