+++ actblue-contributions/blocks/custom/actblue-embed/urlWithQueryConfiguration.js
@@ -1,0 +1,14 @@
+export default ({ url, refcode }) => {
+	const queryParams = [];
+
+	if (refcode) {
+		queryParams.push(`refcode=${refcode}`);
+	}
+
+	const queryString = queryParams.length ? queryParams.join("&") : false;
+
+	if (url && queryString) {
+		return `${url}?${queryString}`;
+	}
+	return url;
+};
